Integrate LLVM at llvm/llvm-project@6a8ba31
Updates LLVM usage to match llvm/llvm-project@6a8ba31.
Further updates the MLIR-HLO submodule to tensorflow/mlir-hlo@ef42c50.

Upstream split `std.splat` into `tensor.splat` and `vector.splat` with
commit llvm/llvm-project@6a8ba31. Hence, the tensor-to-emitc pass now
has to support converting `tensor.splat` to an `emitc.call`. The
corresponding reference implementation is moved from `emitc::standard`
to `emitc::tensor`. Furthermore, since no other std operations are
supported via conversions, all std related passes are cleaned out.
