+++ docs/std-op-coverage.md
@@ -9,4 +9,3 @@ The table below shows the supported Standard ops.
 | cond_br               | :white_check_mark: | via `emitc-translate` |
 | constant              | :white_check_mark: | via `emitc-translate` |
 | return                | :white_check_mark: | via `emitc-translate` |
-| splat                 | :heavy_check_mark: | |
