+++ docs/tensor-op-coverage.md
@@ -5,3 +5,4 @@ The table below shows the supported Tensor ops.
 | op                    | supported          | comment |
 | :-------------------- |:------------------:| :------ |
 | extract               | :heavy_check_mark: | |
+| splat                 | :heavy_check_mark: | |
