+++ include/emitc/InitPasses.h
@@ -41,15 +41,12 @@ inline void registerAllEmitCPasses() {
   registerMHLOToEmitCPipeline();
 #endif // EMITC_BUILD_HLO
   registerConvertArithToEmitCPass();
-  registerConvertStdToEmitCPass();
   registerConvertTensorToEmitCPass();
   registerConvertTosaToEmitCPass();
   registerInsertEmitCArithIncludePass();
-  registerInsertEmitCStdIncludePass();
   registerInsertEmitCTensorIncludePass();
   registerInsertEmitCTosaIncludePass();
   registerArithToEmitCPipeline();
-  registerStdToEmitCPipeline();
   registerTensorToEmitCPipeline();
   registerTosaToEmitCPipeline();
 }
