+++ reference-implementation/include/emitc/tosa.h
@@ -18,7 +18,7 @@
 #include <limits>
 
 #include "emitc/core_ops.h"
-#include "emitc/std.h"
+#include "emitc/tensor.h"
 
 #ifdef EMITC_TOSA_USE_EIGEN
 #include "emitc/tosa_eigen.h"
@@ -572,7 +572,7 @@ Dest slice(Src x, Tensor<int64_t, Src::rank()> start_indices,
   Tensor<int64_t, Src::rank()> limit_indices =
       emitc::add(start_indices, slice_sizes);
   Tensor<int64_t, Src::rank()> strides =
-      emitc::standard::splat<Tensor<int64_t, Src::rank()>>(1);
+      emitc::tensor::splat<Tensor<int64_t, Src::rank()>>(1);
   return emitc::slice<Dest, Src>(x, start_indices, limit_indices, strides);
 }
